数据库调整：对金额的设置统一修改为float

+++ ChunChen-CRM/ChunChen-CRM.Model/Model/CustomerDetailModel.cs
@@ -71,7 +71,7 @@ namespace ChunChen_CRM.Model
         /// <summary>
         /// 消费总金额
         /// </summary>
-        public int Spend { get; set; }
+        public double Spend { get; set; }
 
         /// <summary>
         /// 消费报表信息
